requirements.txt
pytube==15.0.0
xrpl-py==1.8.0
requests==2.31.0
Django==4.2.2
psycopg2==2.9.6
djangorestframework==3.14.0
redis==4.5.5
drf-yasg==1.21.5
sib-api-v3-sdk==7.6.0
djangorestframework-simplejwt==5.2.2